Moved SkyBox to it's own component
